% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{add_convo_stub}
\alias{add_convo_stub}
\title{Add new stub to convo object}
\usage{
add_convo_stub(convo, level, stub, desc = NA, valid = NA)
}
\arguments{
\item{convo}{A \code{convo} object}

\item{level}{Level at which to add new entry}

\item{stub}{Character. New stub to add to controlled vocabularly at given level}

\item{desc}{Character. Optional human-readable description of stub}

\item{valid}{Character vector. Optional \code{pointblank} validation checks for stub}
}
\value{

}
\description{
This function adds new stubs to a controlled vocabularly or overwrites existing stubs.
}
\examples{
stubs <- list(letters[1:3], letters[4:5], letters[6:7])
convo <- create_convo(stubs)
add_convo_stub(convo, 1, "z")
}
