% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_convo}
\alias{compare_convo}
\title{Compare vocabularies with set operations}
\usage{
compare_convo(vocab1, vocab2, fx = c("setdiff", "intersect", "union"))
}
\arguments{
\item{vocab1}{Vocabulary list (or \code{convo} object). In the case of validating a new vocabulary
against an existing one, this argument should take the new vocabulary.}

\item{vocab2}{Vocabulary list (or \code{convo} object). In the case of validating a new vocabulary
against an existing one, this argument should take the existing vocabulary.}

\item{fx}{Character string describing comparison function. \code{setdiff} returns those in \code{vocab1} not in \code{vocab2}; \code{intersect}
returns those in both \code{vocab1} and \code{vocab2}; \code{union} returns all elements in either}
}
\value{
List (of class \code{convomin}) of mismatches
}
\description{
Compare vocabularies with set operations
}
\examples{
v1 <- list(letters[1:3], letters[4:6], letters[7:9])
v2 <- list(letters[(1:3)+1], letters[(4:6)-2], letters[7])
compare_convo(v1, v2, "setdiff")
compare_convo(v1, v2, "union")
compare_convo(v1, v2, "intersect")
}
