% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_convo}
\alias{create_convo}
\title{Create convo object from list of stubs}
\usage{
create_convo(l)
}
\arguments{
\item{l}{List of stubs with one element per level containing a character vector of stubs}
}
\value{
\code{convo} object
}
\description{
Create convo object from list of stubs
}
\examples{
stubs <- list(letters[1:3], letters[4:5], letters[6:7])
create_convo(stubs)
}
