% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointblank.R
\name{create_pb_agent}
\alias{create_pb_agent}
\title{Generate agent object for validation of controlled vocabulary conditions for a data set}
\usage{
create_pb_agent(convo, tbl, level = 1)
}
\arguments{
\item{convo}{\code{convo} object read from YAML}

\item{tbl}{The input table passed to \code{pointblank::create_agent()}}

\item{level}{Numeric. Level of controlled vocabulary containing validation checks}
}
\value{
\code{pointblank} agent for validation pipeline
}
\description{
Generate agent object for validation of controlled vocabulary conditions for a data set
}
\examples{
filepath <- system.file("", "ex-convo.yml", package = "convo")
convo <- read_convo(filepath)
agent <- create_pb_agent(convo, data.frame(IND_A = 1, IND_B = 5, DT_B = as.Date("2020-01-01")))
}
