% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe_names}
\alias{describe_names}
\title{Create dictionary defining controlled vocabulary realizations}
\usage{
describe_names(vars, convo, desc_str = "{level1} of entity {level2}")
}
\arguments{
\item{vars}{Character vector of variable names}

\item{convo}{Controlled vocabulary object}

\item{desc_str}{Prototypical string from which to create variable definition}
}
\value{
\code{data.frame} data dictionary
}
\description{
Create dictionary defining controlled vocabulary realizations
}
\examples{
vars <- c("AMT_A_2019", "IND_B_2020")
filepath <- system.file("", "ex-convo.yml", package = "convo")
convo <- read_convo(filepath)
describe_names(vars, convo, desc_str = "{level1} of {level2} in given year")
}
