% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convo_func.R
\name{DigFilter}
\alias{DigFilter}
\title{Bandpass Digital Filter}
\usage{
DigFilter(x, fs, LD, LU)
}
\arguments{
\item{x}{numeric sequence}

\item{fs}{sampling frequency of a numeric sequence \code{x}}

\item{LD}{lower bandpass frequency}

\item{LU}{higher bandpass frequency}
}
\description{
Bandpass digital filter; passes frequencies within a certain range and
rejects frequencies outside that range.
}
\examples{
## generate components of a signal x
fs <- 1000
t.seq <- seq(0, 1, length.out = 1000)
x1 <- sin(2 * pi * t.seq * 50)
x2 <- sin(2 * pi * t.seq * 150)
x3 <- sin(2 * pi * t.seq * 250)
x0 <- rnorm(length(t.seq), sd = 0.1)
plot(x0, ylim = range(c(x0, x1, x2, x3)), main = "x components")
lines(x1, col = "blue")
lines(x2, col = "red")
lines(x3, col = "green")
## generate signal x
x <- x1 + x2 + x3 + x0
plot(x, main = "x", type = "l")
## use filter with passband frequencies [100 Hz, 200 Hz]
x.filtered <- DigFilter(x, 1000, 100, 200)
plot(x.filtered, type = "l", main = "x filtered with [100 Hz, 200 Hz] frequency bandpass")
plot(x2, type = "l", main = "compare: x2 component of x")

}
