% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate_convo}
\alias{evaluate_convo}
\title{Evaluate a set of names against a controlled vocabulary}
\usage{
evaluate_convo(convo, vbl_names)
}
\arguments{
\item{convo}{Controlled vocabulary object or list of stub names by level}

\item{vbl_names}{Names to evaluate}
}
\value{
Returns \code{convo} object (list) of violating names by level
}
\description{
Evaluate a set of names against a controlled vocabulary
}
\examples{
convo <- list(c("ind"), letters[1:3], c("pre", "post"))
vbl_names <- c("ind_a", "ind_d", "amt_c", "cat_c_pre", "cat_c_post")
evaluate_convo(convo, vbl_names)
}
