% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_decomp}
\alias{parse_decomp}
\title{Decompose name into component parts}
\usage{
parse_decomp(names, sep, dir = c("beg", "end"))
}
\arguments{
\item{names}{Names created by vocabulary as character vector}

\item{sep}{Separator between different levels of name}
}
\description{
Decompose name into component parts
}
\keyword{internal}
