% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_df}
\alias{parse_df}
\title{Parse vocabulary from a set of names and convert to data.frame}
\usage{
parse_df(names, sep = "_")
}
\value{
A data.frame with one row per name and one column per vocabulary level
}
\description{
Parse vocabulary from a set of names and convert to data.frame
}
\examples{
vbl_names <- c("ind_a", "ind_b", "amt_a", "amt_c", "cat_c_pre", "cat_c_post")
parse_df(vbl_names, sep = "_")
}
