% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_stubs}
\alias{parse_stubs}
\title{Parse vocabulary from a set of names}
\usage{
parse_stubs(names, sep = "_")
}
\value{
List of class \code{convomin} to represent controlled vocabulary
}
\description{
Parse vocabulary from a set of names
}
\examples{
vbl_names <- c("ind_a", "ind_b", "amt_a", "amt_c", "cat_c_pre", "cat_c_post")
parse_stubs(vbl_names, sep = "_")
}
