% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot.R
\name{pivot_convo}
\alias{pivot_convo}
\title{Find index of all stubs in a convo or set of stubs}
\usage{
pivot_convo(convo, repeats_only = TRUE)
}
\arguments{
\item{convo}{A \code{convo} object or list of stubs by level}

\item{repeats_only}{Boolean. Whether to return only stubs which exist in multiple levels}
}
\value{
Named list with one element per stubs (of any level) containing a
character vector of all levels at which that stub appears
}
\description{
Pivot convo structure to have one list element per stub each containing a
numeric vector with the level(s) at which that stub appears (as opposed to one
list element per level with entries for each stub).
}
\details{
The main use case for this function is validating that stubs do not appear at
multiple levels. For this reason, the argument \code{repeats_only} defaults to \code{TRUE}
and only stubs existing in multiple levels are returned.
}
\examples{
convo <- list(letters[1:3], letters[2:4], letters[3:6])
pivot_convo(convo)
pivot_convo(convo, repeats_only = FALSE)
}
