% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.convo}
\alias{print.convo}
\title{Pretty print a convo object}
\usage{
\method{print}{convo}(l)
}
\arguments{
\item{l}{A \code{convo} object}
}
\value{
No return - pretty prints convo object
}
\description{
Pretty print a convo object
}
\examples{
l <- c("a_b")
convo <- parse_stubs(l)
print(convo)
}
