% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.convomin}
\alias{print.convomin}
\title{Pretty print a convomin object}
\usage{
\method{print}{convomin}(x, ...)
}
\arguments{
\item{x}{A \code{convomin} object (a multi-level list of stub names)}

\item{...}{Additional arguments passed to function. Currently unused.}
}
\value{
No return - pretty prints convomin object
}
\description{
\code{convomin} objects are returned by \code{compare_convo()} and \code{evaluate_convo()}
}
\examples{
x <- c("a_b")
convo <- parse_stubs(x)
print(convo)
}
