% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_convo}
\alias{read_convo}
\title{Read controlled vocabulary from YAML}
\usage{
read_convo(path)
}
\arguments{
\item{path}{Path to YAML file containing \code{convo} specification}
}
\value{
Controlled vocabulary as \code{convo} (list) object
}
\description{
Read controlled vocabulary from YAML
}
\examples{
path <- system.file("", "ex-convo.yml", package = "convo")
convo <- read_convo(path)
}
