% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{remove_convo_stub}
\alias{remove_convo_stub}
\title{Remove stub from convo}
\usage{
remove_convo_stub(convo, level, stub)
}
\arguments{
\item{convo}{A \code{convo} object}

\item{level}{Level at which to add new entry}

\item{stub}{Character. Stub to remove from controlled vocabulary at given level}
}
\value{
\code{convo} object
}
\description{
Remove stub from convo
}
\examples{
stubs <- list(letters[1:3], letters[4:5], letters[6:7])
convo <- create_convo(stubs)
remove_convo_stub(convo, 1, "a")
}
