% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convo_func.R
\name{RunningCorr}
\alias{RunningCorr}
\title{Fast Running Correlation Computation}
\usage{
RunningCorr(x, y, circular = FALSE)
}
\arguments{
\item{x}{numeric sequence}

\item{y}{numeric sequence, of equal or shorter length than  \code{x} sequence}

\item{circular}{logical; whether  running correlation is computed assuming
circular nature of  \code{x} sequence (see Details)}
}
\value{
numeric sequence
}
\description{
Computes running correlation between two sequences in a fixed width window,
whose length corresponds to the length of the shorter sequence.  Uses
convolution via Fast Fourier Transform.
}
\details{
Computes running correlation between two sequences in a fixed width window.
The length of a window is equal to the shorter of the two sequences (\code{y}), and window
"runs" over the length of longer sequence (\code{x}).

Parameter \code{circular} determines whether \code{x} sequence is assumed to
have a  circular nature. Assume \eqn{l_x} is the length of longer sequence \code{x}, \eqn{l_y} is the length of shorter sequence \code{y}.
* If \code{circular} equals \code{TRUE}, then the output sequence has length equal to the length of \code{x} sequence. First element of the output sequence corresponds to correlation between \code{x[1:l_y]} window of  \code{x} sequence and \code{y} sequence.  Last  element of the output sequence corresponds to correlation between \code{x[(l_x):(l_y - 1)]} window of  \code{x} sequence and \code{y} sequence.
* If \code{circular} equals \code{FALSE}, then the output sequence has length equal to  \eqn{l_x - l_y + 1}. First element of the output sequence corresponds to correlation between \code{x[1:l_y]} window of  \code{x} sequence and \code{y} sequence. Last  element of the output sequence corresponds to correlation between \code{x[(l_x - l_y + 1):l_x]} window of  \code{x} sequence and \code{y} sequence.
}
\examples{
x <- sin(seq(0, 1, length.out = 1000) * 2 * pi * 6)
y <- x[1:100]
out1 <- RunningCorr(x, y, circular = TRUE)
out2 <- RunningCorr(x, y, circular = FALSE)
plot(out1, type = "l"); points(out2, col = "red")

}
