% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{viz_names}
\alias{viz_names}
\title{Visualize name set as tree}
\usage{
viz_names(convo_df)
}
\arguments{
\item{convo_df}{\code{data.frame} with variable names split by level}
}
\description{
Visualize name set as tree
}
\examples{
vbls <- c("a_1_x", "a_2", "a_3", "a_4", "b_1", "b_2")
vbls_df <- parse_df(vbls)
viz_names(vbls_df)
}
