% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_convo}
\alias{write_convo}
\title{Write convo config file template from stubs list}
\usage{
write_convo(convo_min, filename = "convo-config.yml", path = ".")
}
\arguments{
\item{convo_min}{List of stubs with one element per level containing a character vector of stubs}

\item{filename}{Desired filename. Defaults to \code{convo-config.yaml}}

\item{path}{Desired filepath. Defaults to current directory}
}
\value{
Does not return. Call function for side-effect of writing file.
}
\description{
Write convo config file template from stubs list
}
\examples{
\dontrun{
valid_stubs <- list(letters[1:3], letters[4:5], letters[6:7])
write_convo(valid_stubs)
convo <- read_convo("convo-config.yml")
print(convo)
evaluate_convo(convo, c("a_d", "a_e_g", "c_b", "d"))
}
}
