% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointblank.R
\name{write_pb}
\alias{write_pb}
\title{Generate YAML for pointblank validation agent from convo}
\usage{
write_pb(convo, col_names, filename = "convo-validation.yml", path = ".")
}
\arguments{
\item{convo}{\code{convo} object read from YAML}

\item{col_names}{Character vector of column names}

\item{filename}{Name of YAML file to create}

\item{path}{Optional path to directory in which YAML file should be created}
}
\value{
Does not return. Call function for side-effect of writing file.
}
\description{
Generate YAML for pointblank validation agent from convo
}
\examples{
\dontrun{
filepath <- system.file("", "ex-convo.yml", package = "convo")
convo <- read_convo(filepath)
write_pb(convo, c("IND_A", "AMT_B"), filename = "convo-validation.yml", path = ".")
}
}
