% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe_convo}
\alias{describe_convo}
\title{Create schema dictionary}
\usage{
describe_convo(convo)
}
\arguments{
\item{convo}{\code{convo} object containing controlled vocabulary}
}
\value{
\code{data.frame} schema dictionary
}
\description{
Create schema dictionary
}
\examples{
filepath <- system.file("", "ex-convo.yml", package = "convo")
convo <- read_convo(filepath)
describe_convo(convo)
}
