% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.convo}
\alias{print.convo}
\title{Pretty print a convo object}
\usage{
\method{print}{convo}(x, ...)
}
\arguments{
\item{x}{A \code{convo} object}

\item{...}{Additional arguments passed to function. Currently unused.}
}
\value{
No return - pretty prints convo object
}
\description{
Pretty print a convo object
}
\examples{
convo <- create_convo(list(letters[1:3], letters[4:6]))
print(convo)
}
