% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convo_func.R
\name{RunningSd}
\alias{RunningSd}
\title{Fast Running Standard Deviation Computation}
\usage{
RunningSd(x, W, circular = FALSE)
}
\arguments{
\item{x}{numeric sequence}

\item{W}{numeric; width of \code{x} sequence window}

\item{circular}{logical; whether  running sample standard deviation is computed assuming
circular nature of  \code{x} sequence (see Details)}
}
\value{
numeric sequence
}
\description{
Computes running sample standard deviation of a sequence in a fixed width window. Uses
convolution via Fast Fourier Transform.
}
\details{
Parameter \code{circular} determines whether \code{x} sequence is assumed to have a  circular nature. Assume \eqn{l_x} is the length of sequence \code{x}, \code{W} is a fixed length of code{x} sequence window.
* If \code{circular} equals \code{TRUE}, then the output sequence has length equal to the length of \code{x} sequence. First element of the output sequence corresponds to sample standard deviation of \code{x[1:W]} window of  \code{x} sequence.  Last  element of the output sequence corresponds to sample standard deviation of \code{x[(l_x):(W - 1)]} window of  \code{x} sequence.
* If \code{circular} equals \code{FALSE}, then the output sequence has length equal to  \eqn{l_x - W + 1}. First element of the output sequence corresponds to sample standard deviation of  \code{x[1:W]} window of  \code{x} sequence. Last  element of the output sequence corresponds to sample standard deviation of \code{x[(l_x - W + 1):l_x]} window of  \code{x} sequence.
}
\examples{
x <- rnorm(1000)
RunningSd(x, 100)
length(RunningSd(x, 100, circular = FALSE))
length(RunningSd(x, 100, circular = TRUE))

}
